requirements.txt
numpy
pandas
scikit-learn
toml
pyarrow
plotly
evidently
streamlit
Pillow
altair<5